requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async runtime
asyncio==3.4.3
aiohttp==3.9.1
aiodns==3.1.1

# API clients
anthropic==0.8.1
alpaca-py==0.14.0

# RSS parsing
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.4
pyarrow==14.0.1
numpy==1.26.2

# Market data calculations
ta==0.11.0

# Configuration
PyYAML==6.0.1

# Database
aiosqlite==0.19.0

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Scheduling
APScheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
hashlib-additional==1.0.0
